+++ src/saveload/bridge_signal_sl.cpp
@@ -37,7 +37,7 @@ static void Load_XBSS()
 static void RealSave_XBSS(const LongBridgeSignalStorage *lbss)
 {
 	LongBridgeSignalStorageStub stub;
-	stub.length = lbss->signal_red_bits.size();
+	stub.length = (uint32)lbss->signal_red_bits.size();
 	SlObject(&stub, _long_bridge_signal_storage_stub_desc);
 	SlArray(const_cast<uint64*>(&(lbss->signal_red_bits[0])), stub.length, SLE_UINT64);
 }
